# backend/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Plant": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"areas": {
						"items": {
							"$ref": "#/components/schemas/Area"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"address",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Area": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"locationDescription": {
						"type": "string"
					},
					"plant": {
						"$ref": "#/components/schemas/Plant"
					},
					"plantId": {
						"type": "string"
					},
					"equipment": {
						"items": {
							"$ref": "#/components/schemas/Equipment"
						},
						"type": "array"
					},
					"neighbors": {
						"items": {
							"$ref": "#/components/schemas/Area"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"locationDescription",
					"plantId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PartType": {
				"enum": [
					"electric",
					"electronic",
					"mechanical",
					"hydraulical"
				],
				"type": "string"
			},
			"Equipment": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"manufacturer": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					},
					"initialOperationsDate": {
						"type": "string",
						"format": "date-time"
					},
					"area": {
						"$ref": "#/components/schemas/Area"
					},
					"areaId": {
						"type": "string"
					},
					"areas": {
						"items": {
							"$ref": "#/components/schemas/Area"
						},
						"type": "array"
					},
					"parts": {
						"items": {
							"$ref": "#/components/schemas/Part"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"manufacturer",
					"serialNumber",
					"initialOperationsDate",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Part": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/PartType"
					},
					"manufacturer": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					},
					"installationDate": {
						"type": "string",
						"format": "date-time"
					},
					"equipment": {
						"$ref": "#/components/schemas/Equipment"
					},
					"equipmentId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"manufacturer",
					"serialNumber",
					"installationDate",
					"equipmentId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MaintenanceStatus": {
				"enum": [
					"scheduled",
					"completed",
					"overdue",
					"canceled"
				],
				"type": "string"
			},
			"MaintenanceFrequency": {
				"enum": [
					"daily",
					"weekly",
					"monthly",
					"quarterly",
					"semi-annual",
					"annual",
					"custom"
				],
				"type": "string"
			},
			"MaintenanceReferenceType": {
				"enum": [
					"part-installation",
					"equipment-operation",
					"fixed-date"
				],
				"type": "string"
			},
			"Maintenance": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/MaintenanceStatus"
					},
					"frequency": {
						"$ref": "#/components/schemas/MaintenanceFrequency"
					},
					"customDays": {
						"type": "number",
						"format": "double"
					},
					"referenceType": {
						"$ref": "#/components/schemas/MaintenanceReferenceType"
					},
					"part": {
						"$ref": "#/components/schemas/Part"
					},
					"partId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"dueDate",
					"status",
					"frequency",
					"referenceType",
					"partId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Maintenance.Exclude_keyofMaintenance.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/MaintenanceStatus"
					},
					"frequency": {
						"$ref": "#/components/schemas/MaintenanceFrequency"
					},
					"customDays": {
						"type": "number",
						"format": "double"
					},
					"referenceType": {
						"$ref": "#/components/schemas/MaintenanceReferenceType"
					},
					"part": {
						"$ref": "#/components/schemas/Part"
					},
					"partId": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"dueDate",
					"status",
					"frequency",
					"referenceType",
					"partId"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Maintenance.id-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_Maintenance.Exclude_keyofMaintenance.id-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_Omit_Maintenance.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"dueDate": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"$ref": "#/components/schemas/MaintenanceStatus"
					},
					"frequency": {
						"$ref": "#/components/schemas/MaintenanceFrequency"
					},
					"customDays": {
						"type": "number",
						"format": "double"
					},
					"referenceType": {
						"$ref": "#/components/schemas/MaintenanceReferenceType"
					},
					"part": {
						"$ref": "#/components/schemas/Part"
					},
					"partId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/maintenance": {
			"get": {
				"operationId": "GetMaintenances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Maintenance"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMaintenance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Maintenance"
								}
							}
						}
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_Maintenance.id-or-createdAt-or-updatedAt_"
							}
						}
					}
				}
			}
		},
...
		"/maintenance/{maintenanceId}": {
			"get": {
				"operationId": "GetMaintenance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Maintenance"
								}
							}
						}
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "maintenanceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMaintenance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Maintenance"
								}
							}
						}
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "maintenanceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Omit_Maintenance.id-or-createdAt-or-updatedAt__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMaintenance",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "maintenanceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/maintenance/part/{partId}": {
			"get": {
				"operationId": "GetMaintenancesByPart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Maintenance"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "partId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/maintenance/{maintenanceId}/status": {
			"put": {
				"operationId": "UpdateMaintenanceStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Maintenance"
								}
							}
						}
					}
				},
				"tags": [
					"Maintenance"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "maintenanceId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"status": {
										"$ref": "#/components/schemas/MaintenanceStatus"
									}
								},
								"required": [
									"status"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}